Introduced OQ_SAMPLES_SOURCES [skip hazardlib]
