+++ openquake/calculators/base.py
@@ -396,7 +396,7 @@ class HazardCalculator(BaseCalculator):
                     self.csm = self.csm.grp_by_src()
             with self.monitor('splitting sources', measuremem=1, autoflush=1):
                 logging.info('Splitting sources')
-                self.csm.split_all()
+                self.csm.split_all(os.environ.get('OQ_SAMPLE_SOURCES'))
             if self.is_stochastic:
                 # initialize the rupture serial numbers before filtering; in
                 # this way the serials are independent from the site collection
